refactor: 메모 @Focus 기능 리팩토링
