+++ DiverBook_iOS/DiverBook_iOS/Sources/Presentation/DiverProfileScene/View/DiverProfileView.swift
@@ -88,38 +88,16 @@ struct DiverProfileView: View {
                     }
                     .animation(.easeOut(duration: 0.1), value: keyboardObserver.keyboardHeight)
                 }
-                .onChange(of: viewModel.memo) { newValue in
-                    viewModel.action(.memoChanged(newValue))
-                }
             }
         }
         .padding(.horizontal, 24)
         .onAppear {
             viewModel.action(.viewAppeared)
         }
+        .onChange(of: viewModel.memo) { newValue in
+            viewModel.action(.memoChanged(newValue))
+        }
         .hideKeyboardOnTap()
     }
 }
 
-import SwiftUI
-import Combine
-
-final class KeyboardObserver: ObservableObject {
-    @Published var keyboardHeight: CGFloat = 0
-
-    private var cancellables = Set<AnyCancellable>()
-
-    init() {
-        let willShow = NotificationCenter.default.publisher(for: UIResponder.keyboardWillShowNotification)
-            .compactMap { $0.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? CGRect }
-            .map { $0.height }
-
-        let willHide = NotificationCenter.default.publisher(for: UIResponder.keyboardWillHideNotification)
-            .map { _ in CGFloat(0) }
-
-        Publishers.Merge(willShow, willHide)
-            .receive(on: RunLoop.main)
-            .assign(to: \.keyboardHeight, on: self)
-            .store(in: &cancellables)
-    }
-}
